Fixed PLSZ debug issue and 68K reset instruction which should fix Outrun.

+++ yabause/src/musashi/m68kmake.c
@@ -588,7 +588,7 @@ int fgetline(char* buff, int nchars, FILE* file)
 	if(buff[0] == '\r')
 		memcpy(buff, buff + 1, nchars - 1);
 
-	length = strlen(buff);
+	length = (int)strlen(buff);
 	while(length && (buff[length-1] == '\r' || buff[length-1] == '\n'))
 		length--;
 	buff[length] = 0;
